# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Passwordless Multi-Auth API",
    "description": "Passwordless Multi-Auth Platform API - Passwordless Authentication\n\nThis API provides comprehensive endpoints for:\n- MagicLink-based passwordless authentication\n- User management and security operations\n- Role and authority management\n- Push notifications\n- TOTP/MFA verification\n\n## Authentication Flow\n1. Call `POST /api/auth/email-magiclink/send` with user email\n2. User opens the emailed link in a browser: `GET /api/auth/verify?token={token}`\n3. The server redirects (302) to the app scheme `passwordless://auth/callback?code={code}` (or `?status=totp_required` when MFA is enabled, `?status=error&reason=...` on failure)\n4. Trade the one-time code for a token pair via `POST /api/auth/exchange` (if TOTP was required, call `POST /api/auth/totp/verify` instead)\n5. Use the returned JWT in the Authorization header: `Bearer {token}`\n6. Renew with `POST /api/auth/refresh` (single-use rotation); revoke with `POST /api/auth/logout`\n\n## Flutter Integration\nFor Flutter apps, use the `/api/auth/email-magiclink/send` endpoint\nfor mobile authentication. The MagicLink can be handled via\ndeep linking in your Flutter app.\n",
    "contact": {
      "name": "Passwordless Multi-Auth Team",
      "email": "api@example.com"
    },
    "license": {
      "name": "Private",
      "url": "https://github.com/ersincivi/spring-boot-passwordless-multi-auth"
    },
    "version": "1.0.0"
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/users/{username}/security": {
      "get": {
        "tags": [
          "User Management"
        ],
        "operationId": "getUserSecurity",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserSecurityResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/{username}/profile": {
      "get": {
        "tags": [
          "User Management"
        ],
        "summary": "Get user profile",
        "description": "Get user profile information (optimized projection). Users can access their own profile, admins can access any profile.",
        "operationId": "getUserProfile",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "Username of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUserProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/oauth": {
      "get": {
        "tags": [
          "User Management"
        ],
        "operationId": "getOAuthUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListUserSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/locked": {
      "get": {
        "tags": [
          "User Management"
        ],
        "operationId": "getLockedUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListUserSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/inactive": {
      "get": {
        "tags": [
          "User Management"
        ],
        "operationId": "getInactiveUsers",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 90
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListUserSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/users/active": {
      "get": {
        "tags": [
          "User Management"
        ],
        "operationId": "getActiveUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListUserSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles": {
      "get": {
        "tags": [
          "Role Management"
        ],
        "summary": "Get all roles",
        "description": "Get all roles with authorities for management. Performance: ~85% less data, optimized for batch operations. Security: Admin only access.",
        "operationId": "getAllRoles",
        "responses": {
          "200": {
            "description": "Roles retrieved successfully",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/{code}": {
      "get": {
        "tags": [
          "Role Management"
        ],
        "summary": "Get role by code",
        "description": "Get role with authorities - optimized for administrative operations. Performance: ~60% less data with optimized JOINs. Security: Admin only access.",
        "operationId": "getRole",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Role code (USER, ADMIN, SERVICE)",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "USER",
                "SERVICE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role retrieved successfully",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "Role not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRoleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/roles/basic": {
      "get": {
        "tags": [
          "Role Management"
        ],
        "summary": "Get basic roles",
        "description": "Get basic roles for user assignment. Performance: Lightweight projection for role selection. Security: Admin only access.",
        "operationId": "getBasicRoles",
        "responses": {
          "200": {
            "description": "Basic roles retrieved successfully",
            "content": {
              "application/json": {}
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseUserSecurityResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserSecurityResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSecurityResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean"
          },
          "mfaEnabled": {
            "type": "boolean"
          },
          "phoneNumber": {
            "type": "string"
          },
          "oauthProvider": {
            "type": "string"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginIp": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "locale": {
            "type": "string"
          }
        }
      },
      "ApiResponseUserProfileResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserProfileResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "lastLoginIp": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          }
        }
      },
      "ApiResponseListUserSummaryResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummaryResponse"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSummaryResponse": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "locked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "locale": {
            "type": "string"
          }
        }
      },
      "ApiResponseRoleResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RoleResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "SERVICE"
            ]
          },
          "name": {
            "type": "string"
          },
          "authorities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          },
          "locale": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Enter JWT token obtained from MagicLink authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}